+++ mpx_dashboard/plotting.py
@@ -1,468 +1,19 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 """
-Created on Mon Sep 12 19:10:04 2022
+Created on Wed Oct 26 17:59:34 2022
 
-@author: nico
+@author: nr
 """
-
+from typing import Optional, Union#, Any
 import streamlit as st
 import pandas as pd
 import numpy as np
 import plotly.graph_objects as go
 import plotly.express as px
-from typing import Optional, Union, Any
 from mpx_dashboard import csv_specs 
-import pyarrow.parquet as pq
-import pyarrow as pa
+from mpx_dashboard.postprocess import pad_dates, aggr_groups
 from mpx_dashboard import utils
-from epigraphhub.settings import env
-from epigraphhub.connection import get_engine
-import wbgapi as wb
-
-def cols_to_dict(df:pd.DataFrame, a: str, b: str) -> dict :
-    """
-    Obtains a dictionary based on correspondence between two columns, for instance country names and ISO3 codes
-    
-    Parameters
-    ----------
-    df : pd.DataFrame
-        the dataframe to start from
-    a : str
-        the column for dictionary keys
-    b : str
-        the column for dictionary values
-    
-    Returns
-    -------
-    dict
-        dictionary of corrispondence of values in 'a' to values in 'b'
-    """
-    return df.set_index(a)[b].to_dict()
-
-is_bijective: lambda x: len(set(x.keys())) == len(set(x.values()))  # whether dictionary is bijective/reversible
-
-def keys_same_val(dict_: dict) -> dict:
-    """
-    Finds all keys in a dictionary that have the same values. 
-    Useful to check if a dictionary can be inverted, or make it invertable.
-    
-    Parameters
-    ----------
-    dict_ : dict
-        the dictionary to analyse.
-
-    Returns
-    -------
-    to_return : dict
-        {value: [list of keys with that value]}
-
-    """
-    s = pd.Series(dict_)
-    dup_vals = list(set(s[s.duplicated()]))
-    to_return = {}
-    for dup in dup_vals:
-        to_return[dup] = s[s == dup].index.tolist()
-    return to_return
-    
-def get_locality_names() -> pd.DataFrame:
-    """
-    Returns
-    -------
-    pd.DataFrame
-        description of the location_key values (google API)
-
-    """
-    return pd.read_sql_table('locality_names_0',
-                             get_engine(env.db.default_credential),
-                             schema = 'google_health')
-
-def location_name_to_location_key(loc: str, loc_table: Optional[pd.DataFrame] = None) -> str:
-    """
-    Tries to convert location name to location key, starting from countries and going to smaller entities.
-    N. B. it may have issues due to different spelling or language (Cabo Verde/Cape Verde) and multiple places with same name
-    (cities named as regions, etc...). Consider as helpful but not infallible.
-    
-    Parameters
-    ----------
-    loc : str
-        location name.
-
-    Returns
-    -------
-    str
-        location id (as in google API).
-
-    """
-    if loc_table is None:
-        loc_table = get_locality_names()
-    locs0 = loc_table[loc_table['aggregation_level'] == 0]
-    if loc in locs0['country_name'].values:
-        return locs0[locs0['country_name'] == loc]['location_key'].iloc[0]
-    locs1 = loc_table[loc_table['aggregation_level'] == 1]
-    if loc in locs1['subregion1_name'].values:
-        return locs1[locs1['subregion1_name'] == loc]['location_key'].iloc[0]
-    locs2 = loc_table[loc_table['aggregation_level'] == 2]
-    if loc in locs2['subregion2_name'].values:
-        return locs2[locs2['subregion2_name'] == loc]['location_key'].iloc[0]
-    locs3 = loc_table[loc_table['aggregation_level'] == 3]
-    if loc in locs3['locality_name'].values:
-        return locs3[locs3['locality_name'] == loc]['location_key'].iloc[0]
-    
-def get_iso2_to_iso3() -> dict:
-    """
-    Legacy, now this dictionary is available in utils
-
-    Returns
-    -------
-    dict
-        dictionary of iso2 to iso3.
-
-    """
-    df = get_locality_names()
-    return cols_to_dict(df, 'iso_3166_1_alpha_2', 'iso_3166_1_alpha_3')
-
-@st.cache(allow_output_mutation=True)
-def get_demographics_egh() -> pd.DataFrame:
-    """
-
-    Returns
-    -------
-    pd.Dataframe
-        the demographics table from the google API
-
-    """
-    return pd.read_sql_table('demographics', 
-                             get_engine(env.db.default_credential),
-                             schema = 'google_health')
-
-def get_country_pop_egh(loclist: Optional[list] = None, type_: str = 'country') -> pd.Series:
-    """
-    Gets the population for all countries required, from google API
-    
-    Parameters
-    ----------
-    loclist : Optional[list], optional
-        the list of locations desired. The default is None.
-    type_ : str, optional
-        the type of location specifications. Default is 'country'. 
-        Options are 'country' (case insensitive) anything containing 'iso3' (case insensitive), 
-        otherwise 'location_key' is supposed. For countries, location_key is the iso2 code.
-
-    Returns
-    -------
-    pd.Series
-        the total population for each country in loclist, index is as loclist, according to type_,
-        from the google API
-
-    """
-    pop = get_demographics_egh()
-    if loclist:
-        corr = False
-        if type_.lower() == 'country':
-            corr = {utils.country_to_location_key[i]: i for i in loclist}
-            loclist = list(corr.keys())
-        elif 'iso3' in type_.lower():
-            corr = utils.iso2_to_iso3
-            loclist = [utils.iso3_to_iso2[i] for i in loclist]
-        pop = pop[pop['location_key'].isin(loclist)]
-    else:
-        corr = utils.location_key_to_country if type_.lower() == 'country' else utils.iso3_to_iso2 if 'iso2' in type_.lower() else False
-    pop[type_] = pop['location_key'].map(corr) if corr else pop['location_key']
-    return pop.set_index(type_)['population']
-
-@st.cache(allow_output_mutation=True)
-def get_wb_data(time: str, ind: Union[str, list] = ["SP.POP.TOTL.FE.IN", "SP.POP.TOTL.MA.IN"],
-                country: Union[list, str] = 'all') -> pd.DataFrame:
-    """
-    Parameters
-    ----------
-    time : str
-        time specification (e.g. YR2021).
-    ind : Union[str, list], optional
-        Field(s) required. The default is ["SP.POP.TOTL.FE.IN", "SP.POP.TOTL.MA.IN"].
-    country : Union[list, str], optional
-        ISO3 of the country/ies desired. The default is 'all'.
-
-    Returns
-    -------
-    pd.DataFrame
-        the requested table from the world bank.
-
-    """
-    return wb.data.DataFrame(series=ind, economy=country, db=2, time=time)
-    
-def get_country_pop_wb(year: Optional[int] = None,
-                    loclist: Optional[list] = None, type_: str = 'country') -> pd.Series:
-    """
-    Parameters
-    ----------
-    year : int, optional
-        The year that you want to extract the data. Default is None and uses most recent available
-    loclist : list, optional
-        the list of locations desired
-    type_ : str
-        the type of location specifications. Default is 'country'. 
-        Options are 'country' (case insensitive) anything containing 'iso2' (case insensitive), 
-        otherwise 'iso3' is supposed
-    Returns
-    -------
-    pd.Series
-        the total population for each country in loclist, index is as loclist, according to type_,
-        from the world bank.
-    """
-    if year == None:
-        year = pd.to_datetime('today').year
-
-    ind = ["SP.POP.TOTL.FE.IN", "SP.POP.TOTL.MA.IN"] 
-    if loclist:
-        corr = False
-        if type_.lower() == 'country':
-            corr = {utils.country_to_iso3[i]: i for i in loclist}
-            loclist = list(corr.keys())
-        elif 'iso2' in type_.lower():
-            corr = utils.iso3_to_iso2
-            loclist = [utils.iso2_to_iso3[i] for i in loclist]
-    else:
-        corr = utils.iso3_to_country if type_.lower() == 'country' else utils.iso3_to_iso2 if 'iso2' in type_.lower() else False
-    while True:
-        try:
-            time = f'YR{year}' # specify the year that you want to get 
-            df = get_wb_data(ind, time, country=loclist)
-            break
-        except wb.APIResponseError: 
-            year -= 1
-    df = df.reset_index()
-    df = df.rename(columns = {'economy': 'country', 'SP.POP.TOTL.FE.IN': 'total_female_pop', 'SP.POP.TOTL.MA.IN': 'total_male_pop'})
-    df[type_] = df['country'].map(corr) if corr else df['country']
-    df.set_index(type_, inplace = True)
-    return df['total_female_pop'] + df['total_male_pop']
-
-@st.cache(allow_output_mutation=True)
-def load_cases(file: str, skiprows: Optional[int] = None, nrows: Optional[int] = None,
-               usecols: Optional[list] = None, maptobool: Optional[list] = None, **kwargs) -> pd.DataFrame:
-    """
-    Reads disaggregated (case by case) data, dealing with type.
-    
-    Parameters
-    ----------
-    file : str
-        the csv to read
-    skiprows : int, optional
-        the number of rows to skip
-    nrows : int, optional
-        number of rows to load. The default is None, i.e. all rows.
-    usecols : list, optional
-        columns to read. The default is None, i.e. all rows.
-    maptobool : list, optional
-        columns to map to a "pseudoboolean" (True, False, NaN)
-    kwargs : dict
-        additional kwargs to pass to read_csv. e.g. dtype
-    Returns
-    -------
-    cases : pd.DataFrame
-        the DataFrame of desired rows and columns, dates are turned to pd.datetime.
-    """
-    if 'parse_date' not in kwargs:
-        if usecols == None:
-            usecols = pd.read_csv(file, nrows=0).columns
-        kwargs['parse_dates'] =  csv_specs.datecols if csv_specs.datecols else [c for c in usecols if csv_specs.isdatecol(c)]
-    if 'dtype' not in kwargs:
-        if usecols is None:
-            usecols = pd.read_csv(file, nrows=0).columns
-        kwargs['dtype'] = {c: csv_specs.guess_dtype(c) for c in usecols}
-    if type(skiprows) == int:
-        skiprows = range(1, skiprows+1)
-    cases = pd.read_csv(file, skiprows=skiprows, nrows=nrows, usecols=usecols,
-                        **kwargs)
-    if maptobool is None:
-        maptobool = [c for c in usecols if csv_specs.ispseudobool(c)]
-    for c in maptobool:
-            cases[c] = cases[c].map(csv_specs.pseudobooldict)
-    return cases 
-
-def group_and_aggr(df: pd.DataFrame, column: str = csv_specs.countrycol, date_col: str = csv_specs.confdatecol,
-                   dropna: bool = True, entrytype: str = 'cases',dropzeros: bool = True) -> pd.DataFrame:
-    """
-    Aggregates data based on a specific column (generally country) and a date column (reporting, confirmation, death...)
-    
-    Parameters
-    ----------
-    df : pd.DataFrame
-        Dataframe with all the cases (one per row).
-    column : str, optional
-        The level to group by. The default is csv_specs.countrycol.
-    date_col : str, optional
-        The date to use as index, cases are summed for each day based on this column. The default is csv_specs.confdatecol.
-    dropna : bool, optional
-        Whether to keep Nans. The default is True.
-    entrytype : str, optional
-        The type of entry. The default is 'cases'.
-    dropzeros : bool, optional
-        Whether to drop days on which there are no reported entrytype. The default is True.
-
-    Returns
-    -------
-    agg : pd.DataFrame
-        the aggregated dataframe (reported entries per date per country).
-
-    """
-    daily_col = f'daily_{entrytype}'
-    df = df[df[date_col].notna()].set_index(date_col)
-    count_col = df.columns[0]
-    agg = (df.groupby(column, observed=True, dropna=dropna)[count_col]
-           .resample('D')
-           .count()
-           .rename(daily_col)
-           .reset_index(level=0)
-           )
-    if dropzeros:
-            agg = agg.loc[agg[daily_col] != 0]
-    return agg
-
-def add_to_parquet(df: pd.DataFrame, path: str) -> None:
-    """
-    Appends to parquet file
-    
-    Parameters
-    ----------
-    df : pd.DataFrame
-        Dataframe to save as parquet
-    path : str
-        the path for the file.
-
-    Returns
-    -------
-    None.
-
-    """
-    pq.write_to_dataset(pa.Table.from_pandas(df) , root_path=path)
-    
-@st.cache(allow_output_mutation=True)
-def cached_read_csv(file: str, **kwargs) -> pd.DataFrame:
-    """
-    cached version of pd.read_csv
-
-    Parameters
-    ----------
-    file : str
-        file for pd.read_csv.
-    **kwargs : dict 
-        any argument for pd.read_csv
-
-    Returns
-    -------
-    pd.DataFrame
-        the content of the csv file
-    """
-    return pd.read_csv(file, **kwargs)
-
-@st.cache(allow_output_mutation=True)
-def cached_read_parquet(file, **kwargs) -> pd.DataFrame:
-    """
-    cached version of pd.read_parquet
-
-    Parameters
-    ----------
-    file : str
-        file for pd.read_parquet.
-    **kwargs : dict 
-        any argument for pd.read_parquet
-
-    Returns
-    -------
-    pd.DataFrame
-        the content of the parquet file
-    """
-    return pd.read_parquet(file, **kwargs)
-
-def aggr_groups(df: pd.DataFrame, dates_padded: bool = True, fill_value: Any = 0) -> pd.DataFrame:
-    """
-    Aggregates all groups (but not dates) in a dataframe. e.g. global cases by aggregating countries
-    
-    Parameters
-    ----------
-    df : pd.DataFrame
-        the dataframe aggregated by groups
-    dates_padded : bool, optional
-        Whether the dates should be padded or not in the return value. The default is True.
-    fill_value : Any, optional
-        value for fillna. The default is 0. Use None to keep NaN.
-        Only matters if dates_padded == True
-
-    Returns
-    -------
-    df : pd.DataFrame
-        the dataframe where all the groups have been aggregated together
-
-    """
-    df = df.groupby(level=0).sum(numeric_only=True)
-    if dates_padded:
-        df = df.asfreq('D')
-        if fill_value != None:
-            df = df.fillna(fill_value)  
-    return df
-    
-def pad_dates(df: Union[pd.Series, pd.DataFrame], **kwargs) -> Union[pd.DataFrame, pd.Series]:
-    """
-    avoids gaps in date ranges by resampling daily and putting a desired value as filling NAs
-    Parameters
-    ----------
-    df : pd.Series, pd.DataFrame
-        series of dataframe whose index values are dates
-    **kwargs : dict
-        any argument to pass to fillna (generally value)
-        if no 'value' is provided, value=0 is used.
-
-    Returns
-    -------
-    pd.Series or pd.DataFrame
-        the original data but index values with the missing dates are introduced and filled with value(default is 0)
-
-    """
-    if not kwargs:
-        kwargs = dict(value=0)
-    return df.asfreq('D').fillna(**kwargs)
-
-@st.cache()
-def get_daily_count(df: pd.DataFrame, index_col: str)  -> pd.Series:
-    """
-    Parameters
-    ----------
-    df : pd.DataFrame
-        dataframe with a case per row
-    index_col : str
-        date column to use as index
-
-    Returns
-    -------
-    pd.Series
-        The count of entries for each day.
-    """
-    count_col = df.columns[0]
-    return df.set_index(index_col)[count_col].resample('D').count()
-
-def na_percentage(df: pd.DataFrame) -> pd.DataFrame:
-    """
-    Parameters
-    ----------
-    df : pd.DataFrame
-        the dataframe to analyse
-    
-    Returns
-    -------
-    pd.DataFrame
-        contains the percentage of nan per column, omitting when 0%
-    """
-    
-    count_na = lambda df, colname: len(df[df[colname].isna()])
-    tot = len(df)/100
-    dict_ = {}
-    for col in df.columns:
-        na_count = count_na(df, col)
-        if na_count:
-            dict_[col] =  na_count/tot
-    return pd.DataFrame(pd.Series(dict_, name='percentage_nans'))
 
 def total_weekly_metrics(aggr: pd.DataFrame, aggr_sus: Optional[pd.DataFrame] = None,
                          do_sus: Optional[bool] = None, key='metrics') -> None:
